2020/league_B/KWAILAB2/requirements.txt
numpy==1.16.2
Pillow==5.4.1
picamera==1.13
